Email, csv, filters:date&status, Main ------- Filter by Date Range, Followed by Filter by Status.  'proper order should be first filter by date and then by status' Send emails to rows selected and also It sends Emails by batches. 10 emails per second max.
New invoice
--------
Customer details and facturation address are filled when customer selected
Customer details and facturation address can be edited and are saved after edited.
Tax Retention is added if the customer has one, or its 0.
Date by default is the current day
Due Date by default is a month later of the current day
Data can be downloaded now as CSV